Unbound variable issue fixed.
The issue was that I wanted to be as clean as possible and wnated to
pass every used variables to he function, but I forget to update the new
function signature in the production test cases, therefore the unbound
variable error. I reverted the 'clean' way, and stayed with the bit of
dirtier global variable setup, as it would be easier to use in the test
cases.